LFS-752: Create script to package nightly file export Move NightlyExport into cardiac-rehab-resources module
