+++ cardiac-rehab-resources/clinical-data/pom.xml
@@ -44,31 +44,6 @@
           </instructions>
         </configuration>
       </plugin>
-      <plugin>
-        <groupId>org.apache.sling</groupId>
-        <artifactId>slingstart-maven-plugin</artifactId>
-        <executions>
-          <execution>
-            <id>attach-prov-model</id>
-            <goals>
-              <goal>attach-slingfeature</goal>
-              <goal>attach-modelarchive</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
     </plugins>
   </build>
-
-  <!-- <dependencies>
-    <dependency>
-      <groupId>org.slf4j</groupId>
-      <artifactId>slf4j-api</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.osgi</groupId>
-      <artifactId>org.osgi.service.component.annotations</artifactId>
-      <version>1.4.0</version>
-    </dependency>
-  </dependencies> -->
 </project>
